log: remove log file when read error

+++ app/database/index.js
@@ -1,4 +1,5 @@
 'use strict'
+const removeFile = require('fs').unlinkSync
 const Datastore = require('nedb-promise')
 const join = require('path').join
 const timer = require('../util/timer')
@@ -8,6 +9,11 @@ function createDatabase (dir, name, option) {
   let opt = { filename: join(dir, name), autoload: true }
   Object.assign(opt, option)
   return new Datastore(opt)
+}
+
+function removeDb (dir, name) {
+  let filename = join(dir, name)
+  removeFile(filename)
 }
 
 function initDatabase (option = {}) {
@@ -40,19 +46,37 @@ function initDatabase (option = {}) {
   let tdStr = timer(td)
 
   const proxyDB = createDb('/log/proxyDB', {
-    onload: function () {
+    onload: function (e) {
+      if (e) {
+        console.error('proxyDB: ', e)
+        console.error('the file will be removed now, you may need to restart service')
+        removeDb(dir, '/log/proxyDB')
+        return
+      }
       proxyDB.remove({ time: { $lte: td } }, { multi: true }).catch(function (e) { console.log(e) })
       proxyDB.remove({ time: { $lte: tdStr } }, { multi: true }).catch(function (e) { console.log(e) })
     },
   })
   const errorDB = createDb('/log/error', {
-    onload: function () {
+    onload: function (e) {
+      if (e) {
+        console.error('errordb: ', e)
+        console.error('the file will be removed now, you may need to restart service')
+        removeDb(dir, '/log/error')
+        return
+      }
       errorDB.remove({ time: { $lte: td } }, { multi: true }).catch(function (e) { console.log(e) })
       errorDB.remove({ time: { $lte: tdStr } }, { multi: true }).catch(function (e) { console.log(e) })
     },
   })
   const hisDB = createDb('/log/his', {
-    onload: function () {
+    onload: function (e) {
+      if (e) {
+        console.error('hisDb: ', e)
+        console.error('the file will be removed now, you may need to restart service')
+        removeDb(dir, '/log/his')
+        return
+      }
       hisDB.remove({ time: { $lte: td } }, { multi: true }).catch(function (e) { console.log(e) })
       hisDB.remove({ time: { $lte: tdStr } }, { multi: true }).catch(function (e) { console.log(e) })
     },
